requirements.txt
# Purpose: Define project dependencies for consistent environment setup
openai>=1.107.0
python-dotenv>=1.1.1
